benchmark runs again. This time with preprocessor defined types

+++ src/util/misc.hpp
@@ -8,18 +8,18 @@ template <typename T> std::vector<T> linspace(T start, T end, size_t count)
   T step = (end - start) / count;
   std::vector<T> ret_vals(count);
 
-  for (int i = 0; i < count; i++)
+  for (size_t i = 0; i < count; i++)
   {
     ret_vals[i] = start + i * step;
   }
   return ret_vals;
 }
 
-std::vector<int> logspace(int start, int end, int base, int sample_cnt)
+std::vector<int> logspace(int start, int end, int base, size_t sample_cnt)
 {
   std::vector<int> n_vals;
   double spacing = (end - start) / sample_cnt;
-  for (int i = 0; i < sample_cnt; i++)
+  for (size_t i = 0; i < sample_cnt; i++)
   {
     double exp = start + i * spacing;
     int n = (int)pow(base, exp);
